# ewb-api/target/generated-sources/com/gain/java/knowledge/api/StudentApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (5.4.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.gain.java.knowledge.api;

import com.gain.java.knowledge.api.model.Student;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
@Validated
@Controller
@Tag(name = "Student", description = "the Student API")
public interface StudentApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /add/student : Add a new student.
     *
     * @param correlationId correlationId to track request. (required)
     * @param student A JSON object containing pet information (required)
     * @return Created (status code 201)
     */
    @Operation(
        operationId = "addStudent",
        summary = "Add a new student.",
        tags = { "Student" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Created")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/add/student",
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> addStudent(
        @Parameter(name = "correlation-id", description = "correlationId to track request.", required = true, schema = @Schema(description = "")) @RequestHeader(value = "correlation-id", required = true) String correlationId,
        @Parameter(name = "Student", description = "A JSON object containing pet information", required = true, schema = @Schema(description = "")) @Valid @RequestBody Student student
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /get/students : Get all students.
     * This is a Student service description
     *
     * @param correlationId ID of pet to fetch (required)
     * @return This means its ok. (status code 200)
     */
    @Operation(
        operationId = "getAllStudent",
        summary = "Get all students.",
        tags = { "Student" },
        responses = {
            @ApiResponse(responseCode = "200", description = "This means its ok.", content = @Content(mediaType = "application/json", schema = @Schema(implementation =  Student.class)))
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/get/students",
        produces = { "application/json" }
    )
    default ResponseEntity<List<Student>> getAllStudent(
        @Parameter(name = "correlation-id", description = "ID of pet to fetch", required = true, schema = @Schema(description = "")) @RequestHeader(value = "correlation-id", required = true) String correlationId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"city\" : \"city\", \"name\" : \"name\", \"last_name\" : \"last_name\", \"section\" : \"section\", \"id\" : 0, \"first_name\" : \"first_name\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}